src/backend/requirements.txt
asgiref==3.3.4
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
Django==3.2.4
django-cors-headers==3.7.0
fastai==2.4
fastcore==1.3.20
fastprogress==1.0.0
idna==2.10
imageio==2.9.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
murmurhash==1.0.5
networkx==2.5.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==8.2.0
preshed==3.0.5
pydantic==1.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.6
sqlparse==0.4.1
srsly==2.4.1
thinc==8.0.5
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wasabi==0.8.2
